requirements.txt
dj-database-url==0.5.0
django==3.2.9
gunicorn==20.1.0
joblib==1.1.0
numpy==1.21.4
pandas==1.3.4
pillow==8.4.0
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
setuptools==59.5.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
whitenoise==5.3.0